Moved azCliUtils from artifacts-common to utility-common.
