--- common-npm-packages/artifacts-common/azCliUtils.d.ts
@@ -1,2 +1,0 @@
-export declare function setAzureCloudBasedOnServiceEndpoint(connectedService: string): void
-export declare function loginAzureRM(connectedService: string): Promise<void>
