+++ common-npm-packages/artifacts-common/package.json
@@ -1,6 +1,6 @@
 {
     "name": "azure-pipelines-tasks-artifacts-common",
-    "version": "2.241.1",
+    "version": "2.242.0",
     "description": "Azure Artifacts common code (for new authentication tasks)",
     "scripts": {
         "build": "cd ../build-scripts && npm install && cd ../artifacts-common && node make.js"
@@ -15,7 +15,7 @@
         "@types/fs-extra": "8.0.0",
         "@types/mocha": "^5.2.6",
         "@types/node": "^16.11.39",
-        "azure-devops-node-api": "12.0.0",
+        "azure-devops-node-api": "14.0.1",
         "azure-pipelines-task-lib": "^4.13.0",
         "fs-extra": "8.1.0",
         "semver": "6.3.0"
