+++ common-npm-packages/utility-common/package.json
@@ -1,6 +1,6 @@
 {
     "name": "azure-pipelines-tasks-utility-common",
-    "version": "3.241.0",
+    "version": "3.242.0",
     "description": "Common Library for Azure Rest Calls",
     "repository": {
         "type": "git",
@@ -20,6 +20,8 @@
         "@types/node": "^16.11.39",
         "azure-pipelines-task-lib": "^4.11.0",
         "azure-pipelines-tool-lib": "^2.0.7",
+        "azure-pipelines-tasks-azure-arm-rest": "^3.242.0",
+        "azure-devops-node-api": "14.0.1",
         "js-yaml": "3.13.1",
         "semver": "^5.7.2"
     },
